Avoid exposing new property methods in EclipseClasspath
